requirements.txt
requests
opencv-contrib-python
visca-over-ip
numpy
Pillow
flask
PyYAML
